% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveThisObject.R
\name{saveThisObject}
\alias{saveThisObject}
\title{Take a selection as an object name and insert code to save and later reload that object.}
\usage{
saveThisObject()
}
\description{
Turns \cr\cr
\code{object_with_a_name}\cr\cr
into\cr\cr
\code{
saveRDS(object_with_a_name, file = "object_with_a_name.rds")
object_with_a_name <- readRDS(file = "object_with_a_name.rds")}
}
